refactor: 스와이프 관련 로직 커스텀 훅으로 분리
